Refactoring: new PedestrianDetector class
